+++ src/MooVC.Testing.Modelling.Account.Open/Open.cs
@@ -3,4 +3,7 @@ namespace MooVC.Testing.Modelling.Account.Open;
 using Muify.Service;
 
 [Creational<Opened>]
-public sealed partial record Open(Owner Owner);
+public sealed partial record Open
+{
+    public Owner Owner { get; init; } = Owner.Unspecified;
+}