+++ src/Muify/Service/FeatureConstructorsVisitor.cs
@@ -3,6 +3,9 @@ namespace Muify.Service
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using Microsoft.CodeAnalysis;
+    using Microsoft.CodeAnalysis.CSharp;
+    using Microsoft.CodeAnalysis.CSharp.Syntax;
     using MooVC.Syntax;
     using MooVC.Syntax.CSharp;
     using Mu.Modelling;
@@ -29,7 +32,9 @@ namespace Muify.Service
                 (Name: "Proposed", Type: typeof(DateTimeOffset)),
             };
 
-            if (feature.Value.Type.IsMutational && feature.Value.Mutational.Type.IsTransitional)
+            bool isTransitional = feature.Value.Type.IsMutational && feature.Value.Mutational.Type.IsTransitional;
+
+            if (isTransitional)
             {
                 Symbol identity = feature.Value.Metadata.TargetIdentity.IsUndefined
                     ? feature.Features.Unit.Value.Identity.GetSymbol(feature.Features.Unit.Namespace)
@@ -60,6 +65,9 @@ namespace Muify.Service
                 .For<Record>(record => record
                     .Named(feature.Value.Name)
                     .WithConstructors(constructor => constructor
+                        .WithArguments(isTransitional ? "target: default" : string.Empty)
+                        .WithScope(Scopes.Public))
+                    .WithConstructors(constructor => constructor
                         .AttributedWith(attribute => attribute
                             .Named((Name: "JsonConstructorAttribute", Qualifier: "System.Text.Json.Serialization")))
                         .WithArguments(arguments)
@@ -70,7 +78,20 @@ namespace Muify.Service
                 .ImportReferences(feature.Namespace)
                 .ToSnippet(Configuration.Options);
 
-            yield return new File(content, $"{feature.Value.Name}.ctor");
+            string source = content;
+
+            if (!isTransitional)
+            {
+                // Builder omits an empty public constructor as undefined, so remove its initializer after rendering.
+                CompilationUnitSyntax declaration = SyntaxFactory.ParseCompilationUnit(source);
+                ConstructorDeclarationSyntax constructor = declaration.DescendantNodes()
+                    .OfType<ConstructorDeclarationSyntax>()
+                    .Single(candidate => candidate.ParameterList.Parameters.Count == 0);
+
+                source = declaration.ReplaceNode(constructor, constructor.WithInitializer(null)).ToFullString();
+            }
+
+            yield return new File(source, $"{feature.Value.Name}.ctor");
         }
     }
 }